make print_all and other library functions return io::Result<()>
print_all returns a Result with error type Box<dyn Error + Send + Sync>
and this is awkward to work with and can't be ?-converted to
anyhow::Error because it's the trait object is unsized.

Furthermore, the only type of error that print_all can return is
io::Error, so why not just use that directly?

Make header, footer, and print_position_indicator methods also return
io::Result<()> rather than just hiding all write errors.

Technically this is a breaking change, but this is a fork and I don't
want to have diverging version numbers.