lib: add Printer::with_default_style, own writer rather than reference
Printer::new needing 6 options, 4 of which are bools, is ugly. A builder
pattern would be more proper, but I'm lazy so just add a shortcut to get
the default full-featured style easily.

Also there's no need for Printer to hold an exclusive reference to some
writer, it can just own it. std has `impl<W: Write> Write for &mut W`
anyway, so an exclusive reference to a writer is itself also a writer.