restore colors to hexyl v0.10's appearance
In commit 98ceec1 "Replace colors with ANSI styles" the grey color for
offsets and null bytes was changed from a 256-color pallete selection to
ANSI BrightBlack. This is sensible, but I prefer the dimmer grey because
it pops out less, change it back. owo-colors doesn't support 256-color
numbers, but it does have (compile-time const generic!) RGB colors so
use that to get the same color.

In commit 4fc0c1f "Replace `anstyle` with `owo-colors`" the ASCII Other
color got swapped from Magenta to Green. I think this might have just
been a typo when that change was made, because Magenta is standard and
why would we want less color variety?