optimize: remove hot calls to comparing &[u8]
While profiling, I noticed that <[u8] as PartialEq>::eq was surprisingly
hot in Printer::print_byte and Printer::print_char. The reason for this
is that Printer stores the current color as a &'static [u8] slice and
compares the current color from the new/desired color for every
character printed. That comparison is expensive since A) there's levels
of pointer indirection and B) lengths aren't constant (and probably
also because a generic slice equality has to have vectorized cases for
big slices).

Fix by replacing all references to raw ANSI color slices with a new
ColorType enum that's trivial to compare, and only retrieve the ANSI
bytes when printing the colors.

This produces a surprising ~15% speedup.